Create shelter_locator.py
This script will take a file that has a list of homeless shelters in NYC where some of the latitude and and longitude data is missing. It will then use geocoders to determine the latitude and longitude for the missing shelters and write them to a new file called my_results.txt. The input file has the following format -

Neighborhood Coalition For Shelter
Social Services Organization
50 Broadway #1301, New York, NY 10004
(212)537-5100
lat: 40.7063443
long: -74.0126752

NCS Residence
211 East 81st St, New York, NY 10028
(917)677-4504
lat:
long: