Create BlockChain.py
This creates a BlockChain using Block objects. The chain is represented as a Python list. 

To use this, you would create the initial block. Then you would add new blocks to the chain. You add new blocks to the chain by getting the previous nodes proof and then use that to create a new proof of work. Then you would create a transaction which would be a dictionary and the class will add the transaction to the list of transactions. Finally a new node or block is created from the previous nodes hash, the proof that was created, and the list of transactions and added to our chain. See test_blockchain.py for example on usage.

Some of the methods available in the class are as follows -

create_genesis_block
 - this is the first method that needs to be called when creating 
the blockchain and is similar to how one might create a zeroth node when creating a linked list.

create_new_block
- this method creates a new block and adds it to our chain. it takes as an argument a proof and the previous nodes hash. it then instantiates a new block using the previous nodes hash, the proof, and the chain of transactions and adds it to our chain.

create_new_transaction
- this method adds a new dictionary to our list of dictionaries and increments our last block index